Test: ListService Save(), Update(). Delete() 메소드 테스트
ListService.toDoListSave()
ListService.toDoListUpdate()
ListService.toDoListDelete()
테스트 및 결과 확인

* CardRepository 및 ToDoListRepository updateContent() 반환형 Long -> int로 수정
* Card 및 ToDoList Entity int order의 @Column(name = "orderNum")으로 수정
* CardService 및 ListService @Service , @AllArgsConstructor 추가
* CardReopsitory -> CardRepository로 이름 수정
* ListControllerTest 생성

+++ src/main/java/com/example/wa_trello/repository/ToDoListRepository.java
@@ -9,7 +9,7 @@ public interface ToDoListRepository extends JpaRepository<ToDoList,Long> {
     @Modifying(flushAutomatically = true, clearAutomatically = true) //flush,clear는 혹시 모를 기존의 영속성 컨텍스트에 남아있는 쿼리르 수행하기 위해
     // 벌크성(다건의 update,delete를 하나의 쿼리로 하는 것) 수정 쿼리
     @Query(value = "update ToDoList T SET T.title = ?2 where T.id = ?1")
-    Long updateTitle(Long id, String title);
+    int updateTitle(Long id, String title);
 
     @Modifying
     @Query(value = "update ToDoList T SET T.order = ?2 where T.id = ?1")
